Pycrypt
7.24 LAB Activity: PyCrypt
PyCrypt is a secret language of communication. It has the following rules:

We take the first letter of a word, put it at the end, and append “ay”

The only exception is if the first letter is a vowel, in which case we keep it as it is and append “hay” to the end

All upper case characters are converted to lower case

Any non-alphabetic characters remain unchanged

You need to create 2 functions with the following names & behavior:

pycrypt : A function that takes in a word and encrypts that word using the above said rules

 Example: If input is "boot", then output is "ootbay"
pycrypt_sentence : A function that builds on the pycrypt function but the input parameter is a sentence

 Example: If input is "I am a 4 pointer", then output is "ihay amhay ahay 4 ointerpay"
NOTE: There is only one input function & one output function. Depending on whether you're input is a word or a sentence, the appropriate function is used.